+++ wrappers/winbase_amd64.go
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2014-2017 MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the license is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// +build amd64
+
+package wrappers
+
+import (
+	"syscall"
+	"unsafe"
+)
+
+func VerifyVersionInfo(versionInfo *OSVERSIONINFOEX, typeMask uint32, conditionMask uint64) error {
+	r1, _, e1 := syscall.Syscall(
+		procVerifyVersionInfoW.Addr(),
+		3,
+		uintptr(unsafe.Pointer(versionInfo)),
+		uintptr(typeMask),
+		uintptr(conditionMask))
+	if r1 == 0 {
+		if e1 != ERROR_SUCCESS {
+			return e1
+		} else {
+			return syscall.EINVAL
+		}
+	}
+	return nil
+}
