+++ wrappers/winnt_amd64.go
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2014-2017 MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the license is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// +build amd64
+
+package wrappers
+
+import "syscall"
+
+func VerSetConditionMask(conditionMask uint64, typeBitMask uint32, condition uint8) uint64 {
+	r1, _, _ := syscall.Syscall(
+		procVerSetConditionMask.Addr(),
+		3,
+		uintptr(conditionMask),
+		uintptr(typeBitMask),
+		uintptr(condition))
+	return uint64(r1)
+}
